backend/requirements.txt
flask
flask-socketio
python-socketio
python-engineio
pyserial
